% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_conversion.R
\name{Celsius_to_Fahrenheit}
\alias{Celsius_to_Fahrenheit}
\title{Celsius conversion}
\usage{
Celsius_to_Fahrenheit(celcius)
}
\arguments{
\item{celsius}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp1 <- Celsius_to_Fahrenheit(22);
temp2 <- Celsius_to_Fahrenheit( c(-2, 12, 23) );
}
