% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_conversion.R
\name{Fahrenheit_to_Celsius}
\alias{Fahrenheit_to_Celsius}
\title{Fahrenheit conversion}
\usage{
Fahrenheit_to_Celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- Fahrenheit_to_Celsius(50);
temp2 <- Fahrenheit_to_Celsius( c(50, 63, 23) );
}
